Move betty.plugins.place_type to betty.place_types
