+++ betty/place_types/cemetery.py
