+++ betty/place_types/hamlet.py
