+++ betty/place_types/locality.py
