+++ betty/place_types/village.py
