+++ betty/tests/plugins/asset/raspberry_mint/templates/entity/test_summary__place.py
@@ -1,8 +1,8 @@
 from betty.locale.localize import DEFAULT_LOCALIZER
+from betty.place_types.country import Country
 from betty.plugins.asset_directory.raspberry_mint import RASPBERRY_MINT
 from betty.plugins.entity.enclosure import Enclosure
 from betty.plugins.entity.place import Place
-from betty.plugins.place_type.country import Country
 from betty.test_utils.conftest import AssertTemplateFile
