+++ pyproject.toml
@@ -260,29 +260,29 @@ betty = 'betty.console:main_standalone'
 'webpack-entry-point-loader' = 'betty.plugins.js_resource.webpack_entry_point_loader:WEBPACK_ENTRY_POINT_LOADER'
 
 [project.entry-points.'betty.place_type']
-'borough' = 'betty.plugins.place_type.borough:Borough'
-'building' = 'betty.plugins.place_type.building:Building'
-'cemetery' = 'betty.plugins.place_type.cemetery:Cemetery'
-'city' = 'betty.plugins.place_type.city:City'
-'country' = 'betty.plugins.place_type.country:Country'
-'county' = 'betty.plugins.place_type.county:County'
-'department' = 'betty.plugins.place_type.department:Department'
-'district' = 'betty.plugins.place_type.district:District'
-'farm' = 'betty.plugins.place_type.farm:Farm'
-'hamlet' = 'betty.plugins.place_type.hamlet:Hamlet'
-'locality' = 'betty.plugins.place_type.locality:Locality'
-'municipality' = 'betty.plugins.place_type.municipality:Municipality'
-'neighborhood' = 'betty.plugins.place_type.neighborhood:Neighborhood'
-'number' = 'betty.plugins.place_type.number:Number'
-'parish' = 'betty.plugins.place_type.parish:Parish'
+'borough' = 'betty.place_types.borough:Borough'
+'building' = 'betty.place_types.building:Building'
+'cemetery' = 'betty.place_types.cemetery:Cemetery'
+'city' = 'betty.place_types.city:City'
+'country' = 'betty.place_types.country:Country'
+'county' = 'betty.place_types.county:County'
+'department' = 'betty.place_types.department:Department'
+'district' = 'betty.place_types.district:District'
+'farm' = 'betty.place_types.farm:Farm'
+'hamlet' = 'betty.place_types.hamlet:Hamlet'
+'locality' = 'betty.place_types.locality:Locality'
+'municipality' = 'betty.place_types.municipality:Municipality'
+'neighborhood' = 'betty.place_types.neighborhood:Neighborhood'
+'number' = 'betty.place_types.number:Number'
+'parish' = 'betty.place_types.parish:Parish'
 'project-configuration' = 'betty.project._plugin_discovery:_discover_place_types'
-'province' = 'betty.plugins.place_type.province:Province'
-'region' = 'betty.plugins.place_type.region:Region'
-'state' = 'betty.plugins.place_type.state:State'
-'street' = 'betty.plugins.place_type.street:Street'
-'town' = 'betty.plugins.place_type.town:Town'
-'unknown' = 'betty.plugins.place_type.unknown:Unknown'
-'village' = 'betty.plugins.place_type.village:Village'
+'province' = 'betty.place_types.province:Province'
+'region' = 'betty.place_types.region:Region'
+'state' = 'betty.place_types.state:State'
+'street' = 'betty.place_types.street:Street'
+'town' = 'betty.place_types.town:Town'
+'unknown' = 'betty.place_types.unknown:Unknown'
+'village' = 'betty.place_types.village:Village'
 
 [project.entry-points.'betty.role']
 'attendee' = 'betty.roles.attendee:Attendee'
